move logic to api file
